Oauth/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coverage==7.2.7
exceptiongroup==1.1.2
Flask==2.3.2
flask-openapi3==2.4.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pydantic==1.10.12
PyJWT==2.8.0
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
oracledb==1.3.2
pymysql

#pip install openai==0.28.1
#pip install azure-cognitiveservices-speech
#pip install azure-ai-textanalytics==5.2.0
